+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Frontend
 
   - Added db_lock to contract deliverables section
-  
+  - Added db_lock to contract gdx-internal-coding section
 **May 30, 2023**
 
 - Backend